+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * `faspex5`: Support public and private invitations.
     * `faspex5`: Automatic detection of HTTPGW.
     * `faspex5`: Recursive content of package.
+    * `faspex5`: Auto-fill recipient for package send and public link.
   * Issues Fixed:
     * `aoc`: #135 `exclude_dropbox_packages` query option can be overridden
     * **global**: removed gem dependency on `bigdecimal`